move uart0, memory, spl init to boot
