+++ k230_boot/Cargo.toml
@@ -6,3 +6,10 @@ edition = "2024"
 [dependencies]
 pac = { path = "../pac" }
 riscv = "0.12.1"
+
+
+[features]
+default = ["init_ddr", "ddr_2667"]
+init_ddr = []
+ddr_2133 = []
+ddr_2667 = []
