+++ src/main.rs
@@ -37,8 +37,6 @@ macro_rules! print {
 
 pub mod boot;
 pub mod commands;
-#[allow(unused)]
-pub mod ddr_init;
 pub mod readline;
 pub mod serial;
 
@@ -108,81 +106,6 @@ pub fn putc(c: u8) {
     }
 }
 
-unsafe fn spl_device_disable() {
-    // disable ai power
-    if ptr::read_volatile(0x9110302c as *const u32) & 0x2 != 0 {
-        ptr::write_volatile(0x91103028 as *mut u32, 0x30001);
-    }
-
-    // disable vpu power
-    if ptr::read_volatile(0x91103080 as *const u32) & 0x2 != 0 {
-        ptr::write_volatile(0x9110307c as *mut u32, 0x30001);
-    }
-
-    // disable dpu power
-    if ptr::read_volatile(0x9110310c as *const u32) & 0x2 != 0 {
-        ptr::write_volatile(0x91103108 as *mut u32, 0x30001);
-    }
-
-    // disable disp power
-    if ptr::read_volatile(0x91103040 as *const u32) & 0x2 != 0 {
-        ptr::write_volatile(0x9110303c as *mut u32, 0x30001);
-    }
-
-    // check disable status
-    let mut value = 1000000;
-    while (!(ptr::read_volatile(0x9110302c as *const u32) & 0x1 != 0)
-        || !(ptr::read_volatile(0x91103080 as *const u32) & 0x1 != 0)
-        || !(ptr::read_volatile(0x9110310c as *const u32) & 0x1 != 0)
-        || !(ptr::read_volatile(0x91103040 as *const u32) & 0x1 != 0))
-        && value != 0
-    {
-        value -= 1;
-    }
-
-    // disable ai clk
-    value = ptr::read_volatile(0x91100008 as *const u32);
-    value &= !(1 << 0);
-    ptr::write_volatile(0x91100008 as *mut u32, value);
-
-    // disable vpu clk
-    value = ptr::read_volatile(0x9110000c as *const u32);
-    value &= !(1 << 0);
-    ptr::write_volatile(0x9110000c as *mut u32, value);
-
-    // disable dpu clk
-    value = ptr::read_volatile(0x91100070 as *const u32);
-    value &= !(1 << 0);
-    ptr::write_volatile(0x91100070 as *mut u32, value);
-
-    // disable mclk
-    value = ptr::read_volatile(0x9110006c as *const u32);
-    value &= !((1 << 0) | (1 << 1) | (1 << 2));
-    ptr::write_volatile(0x9110006c as *mut u32, value);
-}
-
-unsafe fn board_init() {
-    // UART init
-    // UART is inited by BootROM, so we just disable FIFO and interrupt
-    UART0.ier().write(|w| w.0 = 0);
-    UART0.fcr().write(|w| {
-        w.set_fifoe(false);
-        w.set_xfifor(true);
-        w.set_rfifor(true);
-    });
-
-    println!("\r\n");
-    println!("{}", BANNER);
-    println!("Booting K230 using Rust ....");
-    println!("DDR init ...");
-
-    // spl_board_init_f
-    spl_device_disable();
-    ddr_init::ddr_init_training();
-
-    println!("DDR init done!");
-}
-
 // init UART3
 fn uart_init() {
     let clk_in = 50_000_000;
@@ -232,6 +155,7 @@ fn tsensor_init() {
     });
 }
 
+#[allow(dead_code)]
 fn blinky() {
     // RGB LED of LCKFB
     // - R: GPIO62
@@ -360,9 +284,9 @@ fn shell_repl() {
 entry_point!(main);
 
 fn main(_bootinfo: &'static BootInfo) -> ! {
-    unsafe {
-        board_init();
-    }
+    println!("\r\n");
+    println!("{}", BANNER);
+    println!("Booting K230 using Rust ....");
 
     commands::cpuid();
